Implemented custom type casters for pybind11/SWIG
Using a custom type caster provides a user-friendly way to interface
SWIG-based types with pybind11. The lifetime of the C++ objects is
managed using std::shared_ptr<>. In the case of the SWIG-wrapped object
being created in a code part using pybind11, the Python object is
monkey-patched to include a hidden shared pointer and a custom deleter
destroying the shared pointer attached to the Python object. This, in
turn, results in the destruction of the C++ object if the C++ layer does
not hold any other reference to the shared object.